Refactor project structure
Move ports into separate directory
Remove xml doctype from svg icons
Add keyboard icons
Make styling more BEM-ish
