+++ src/components/layout/editor-button/EditorButton.module.scss
@@ -1,4 +1,3 @@
-//TODO: сделать кнопку отдельным компонентом EditorButton (с опциональной иконкой)
 .editorButton {
   border: none;
   outline: none;
