+++ src/components/layout/editor-button/EditorButton.tsx
@@ -1,0 +1,16 @@
+import React, {HTMLAttributes} from 'react';
+
+import classes from './EditorButton.module.scss';
+
+const EditorButton: React.FC<HTMLAttributes<HTMLButtonElement>> = ({children, className, ...props}) => {
+    return (
+        <button
+            className={`${classes.editorButton} ${className || ''}`}
+            {...props}
+        >
+            {children}
+        </button>
+    );
+};
+
+export default EditorButton;
