+++ src/components/layout/nav-bar/NavigationBar.module.scss
@@ -56,6 +56,8 @@
   display: inline-flex;
   align-items: center;
 
+  text-decoration: none;
+
   padding: 0 16px;
 
   font-family: "Yandex Sans", sans-serif;
