+++ src/components/layout/theme-switcher/ThemeSwitcher.module.scss
@@ -9,5 +9,5 @@
 }
 
 .buttonThemeSwitcher {
-  composes: editorButton from '../button/button.module.scss';
+  composes: editorButton from '../editor-button/EditorButton.module.scss';
 }
