+++ src/components/nodes/processing-node/ProcessingNodeModel.ts
@@ -1,7 +1,7 @@
 import {NodeModel} from '@projectstorm/react-diagrams';
 import {BaseModelOptions, DeserializeEvent} from '@projectstorm/react-canvas-core';
 import {inPortName, outPortName} from '../node/NodeWidget';
-import {FlowPortModel} from '../../links/flow-link/FlowPortModel';
+import {FlowPortModel} from '../../ports/flow-port/FlowPortModel';
 
 export const processingNodeType: string = 'processing-node';
